fix: Also converting unit tests and ask for deleteion
This was an expansive commit ($19,43).
It tried to delete ssh_tests.ssh since it was said to be obsolete (I refused and asked it to convert, which it did).
It actually did a good job on converting all the other resources,
It fixed all linting mistakes, all tests are now running.
It looks like a pretty good prorammer to me!